(refactor): latency store actor placement
